move util functions to shared/utils
