+++ src/uss/actions.ts
@@ -15,8 +15,8 @@ import * as fs from "fs";
 import * as globals from "../globals";
 import * as path from "path";
 import { ZoweUSSNode } from "./ZoweUSSNode";
-import { labelHack, refreshTree, concatChildNodes } from "../shared/utils";
-import { errorHandling, willForceUpload } from "../utils";
+import { labelHack, refreshTree, concatChildNodes, willForceUpload } from "../shared/utils";
+import { errorHandling } from "../utils";
 import { Profiles, ValidProfileEnum } from "../Profiles";
 import { IZoweTree } from "../api/IZoweTree";
 import { IZoweUSSTreeNode } from "../api/IZoweTreeNode";
